+++ src/runner/v3/RunnerV3.ts
@@ -3,7 +3,7 @@ import type { Canvas } from "canvas";
 import type { RunnerStartParameters } from "../Runner";
 import { Runner } from "../Runner";
 import type { RunnerPointEvent } from "../types";
-import { akashicEngine as g, gameDriver as gdr } from "./engineFiles";
+import { akashicEngine as g, gameDriver as gdr, pdi } from "./engineFiles";
 import type { NodeCanvasSurface } from "./platform/graphics/canvas/NodeCanvasSurface";
 import type { NullSurface } from "./platform/graphics/null/NullSurface";
 import { PlatformV3 } from "./platform/PlatformV3";
@@ -148,7 +148,8 @@ export class RunnerV3 extends Runner {
 		this.platform.firePointEvent({
 			type,
 			identifier: event.identifier,
-			offset: event.offset
+			offset: event.offset,
+			button: event.button ?? pdi.PlatformButtonType.Primary
 		});
 	}